+++ src/client/game/electron/preload.ts
@@ -1,0 +1,5 @@
+import { contextBridge } from 'electron';
+
+contextBridge.exposeInMainWorld('mewElectron', {
+  version: process.versions.electron
+});